add additional warnings to the library build and misc clean up switch to ubuntu 24.04 for github build

+++ cpputest-for-qpc-lib/src/cms_cpputest_qf_ctrl.cpp
@@ -68,7 +68,7 @@ void Setup(enum_t const maxPubSubSignalValue, uint32_t ticksPerSecond,
     l_memPoolOption     = memPoolOpt;
     l_ticksPerSecond    = ticksPerSecond;
     l_subscriberStorage = new SubscriberList();
-    l_subscriberStorage->resize(maxPubSubSignalValue);
+    l_subscriberStorage->resize(static_cast<size_t>(maxPubSubSignalValue));
     QSubscrList nullValue;
     QPSet_setEmpty(&nullValue.set);
     QPSet_setEmpty(&nullValue.set_dis);
